chore: fix some stuff, break some stuff
