+++ include/minishell.h
@@ -65,7 +65,6 @@ struct						s_lex
 	int						start;
 	int						end;
 	int						tkn_count;
-	int 						unmatched;
 };
 
 struct						s_shell_data
